#82: rename, add colors for icons, refactor story
