+++ webapp/package.json
@@ -36,6 +36,7 @@
     "autoprefixer": "10.4.20",
     "class-variance-authority": "0.7.0",
     "clsx": "2.1.1",
+    "dayjs": "^1.11.13",
     "lucide-angular": "0.429.0",
     "postcss": "8.4.41",
     "rxjs": "7.8.1",
